Introduce dummy method access flags
One flag marks a method that has been hooked by Xposed. The other one is set on
the backup of the ArtMethod structure, which is used when invoking the original
method.

Hooked methods behave similar to proxy methods and reuse parts of their code.
In some places, updates to the method structure (fixing up trampolines, adding
instrumentation handlers) should be applied to the original (backup) method
instead of the hooked (proxy) method.