Force recompilation of the boot image with Xposed
The boot image needs to be recompiled again to allow hooking framework classes.

The invalidation algorithm needs to ensure that the used image fits to the
original dex/odex files and was compiled for the correct Xposed oat version.
We assume that the system odex is never compiled for Xposed and therefore
just check whether the dalvik-cache version is correct. The checksum will
be different after recompilation, so the original checksum is compared.
If the criteria are not fulfilled, the image will be recompiled.