Fall back to interpreting .odex files when recompiling failed
Zygote and the PackageManagerService ensure that most apps are compiled
before they're used. However, there are some exceptions. For these, the
recompilation is attempted when the app is started, but fails due to
lacking write permissions.

On non-Xposed ROMs, the system falls back to interpreting the .odex file
in such cases. As this doesn't happen very often, the performance impact
is very limited, and the app might actually be compiled later.

This change introduces a similar behavior for Xposed. However, this is
only tried after recompiling the .odex files has failed. The checksums
are ignored in this case, as there isn't much of an alternative anyway -
the app would crash otherwise.