xposed: Change oat 'version' to 'X'.
If the user updates Xposed with a new Xposed release,
it's probably best if a complete recompile is forced.

In particular, the last official 'version' bump was done after the
'Write-protect TLS' commit was reverted; however, our branch doesn't
have that reverted, and I don't have any plans on doing so.